improve ravingcode add makeSHA1u64 fix MultiConfig

+++ src/Kernel/Ravingcode.h
@@ -6,6 +6,6 @@ namespace Mengine
 {
     namespace Helper
     {
-        void ravingcode( uint32_t _parrot, const void * _in, size_t _size, void * const _out );
+        void ravingcode( uint64_t _parrot, const void * _in, size_t _size, void * const _out );
     }
 }
